# docs/apidoc.yaml
openapi: "3.0.0"
info:
  description: "REST API for the Quizzy multiplatform application"
  version: "1.0.0"
  title: "Quizzy REST API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "795306@unizar.es"
  license:
    name: "MIT License"
    url: "https://opensource.org/licenses/MITl"
...
paths:
  /user:
    post:
      tags:
      - "user"
      summary: "Register a new user"
      description: "If there is any user with the same nickname, this operation will fail. Otherwise, a new entry in the user database will be created."
      operationId: "registerUser"
      requestBody:
        $ref: "#/components/requestBodies/UserCredentials"
        
      responses:
        "201":
          description: "user registered successfully"
          content:
            application/json:
              schema: 
                type: "string"
        "409":
          description: "conflict: user already exists"
  /user/login:
    post:
      tags:
        - "user"
      summary: "Logs user into the system."
      description: "If the provided credentials were right, a new session will be opened for this user. Otherwise, this operation will fail."
      requestBody:
        description: "The new user's nickname"
        $ref: "#/components/requestBodies/UserCredentials"
      responses:
        "201":
          description: "user logged in successfully"
          content:
            application/json:
              schema:
                type: "string"
        "409":
          description: "unauthorized: credentials do not match"
...
  /user/{nickname}:
    get:
      tags:
      - "user"
      summary: "Find user by nickname"
      description: > 
        Fetch a user by its nickname. If it exists, the operations returns an User
        object with the desired user's information. Otherwise, the operation fails.
      operationId: "getUserByName"

      parameters:
      - name: "nickname"
        in: "path"
        description: "The nickname of the user that will be fetched."
        required: true
        schema: 
          type: "string"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "404":
          $ref: "#/components/responses/404NotFound"
    put:
      tags:
      - "user"
      summary: "Modify user"
      description: >
        Modifies a user's information.
        
        **The user must be logged in, in order to modify its information.** If 
        the user isn't logged in as the user which is going to be modified, 
        the operation fails.
        
        If the user exists and the provided credentials 
        are correct, the user information will be updated.
      parameters:
      - name: "nickname"
        in: "path"
        required: true
        schema:
          type: "string"
      requestBody:
        $ref: "#/components/requestBodies/UserCredentials"
        
      responses:
        "200":
          description: "Successful operation"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: >
        Deletes a user account and related information.
        
        **The user must be logged in, in order to delete its account.** If 
        the user isn't logged in as the user which is going to be deleted, 
        the operation fails.
        
        If the user exists and the provided credentials 
        are correct, the user account will be deleted from the database.
      parameters:
      - name: "nickname"
        in: "path"
        required: true
        schema:
          type: "string"
      requestBody:
        $ref: "#/components/requestBodies/UserCredentials"
      responses:
        "200":
          description: "Successful operation"
...
components:
  schemas:
    User:
      type: "object"
      properties:
        userId:
          type: "integer"
          format: "uint64"
        nickname:
          type: "string"
          description: "Nickname"
        email:
          type: "string"
          description: "Email"
        password:
          type: "string"
          description: "Encrypted password"
        wallet:
          type: "integer"
          format: "uint64"
          description: "Coin amount in wallet"
        publicWins:
          type: "integer"
          format: "uint64"
        privateWins:
          type: "integer"
          format: "uint64"
        cosmeticId:
          type: "integer"
          format: "uint64"
...
  requestBodies:
    UserCredentials:
      content:
        application/json:
          schema:
            type: "object"
            properties:
              nickname:
                type: "string"
                description: "Nickname"
              email:
                type: "string"
                description: "Email"
              password:
                type: "string"
                description: "Encrypted password"
            required:
            - "nickname"
            - "password"
...
  responses:
    400BadRequest:
      description: Invalid request.
    401Unauthorized:
      description: Access token is missing or invalid.
    404NotFound:
      description: The specified resource was not found.
...